tree-optimization/112444 - avoid bougs PHI value-numbering
With .DEFERRED_INIT ssa_undefined_value_p () can return true for
values we did not visit (because they proved unreachable) but
are not .VN_TOP.  Avoid using those as value which, because they
are not visited, are assumed to be defined outside of the region.

Bootstrapped and tested on x86_64-unknown-linux-gnu, pushed.

	PR tree-optimization/112444
	* tree-ssa-sccvn.cc (visit_phi): Avoid using not visited
	defs as undefined vals.

	* gcc.dg/torture/pr112444.c: New testcase.