amdgcn: Fix vector min/max ICE (pr112313)
I've just committed this patch to fix pr112313 (oops, I forgot to write
the number in the commit message).

The problem was a missed case in the vector reduction expand code.

Andrew
amdgcn: Fix vector min/max ICE

The DImode min/max instructions need a clobber that SImode does not, so
add the special case to the reduction expand code.

gcc/ChangeLog:

	* config/gcn/gcn.cc (gcn_expand_reduc_scalar): Add clobber to DImode
	min/max instructions.