tree-optimization/112281 - loop distribution and zero dependence distances
We currently distribute

  for (c = 2; c; c--)
    for (e = 0; e < 2; e++) {
      d[c] = b = d[c + 1];
      d[c + 1].a = 0;
    }

in a wrong way where the inner loop zero dependence distance should
make us preserve stmt execution order.  We fail to do so since we
only look for a fully zero distance vector rather than looking at
the innermost loop distance.  This is somewhat similar to PR87022
where we instead looked at the outermost loop distance and changed
this to what we do now.  The following switches us to look at the
innermost loop distance.

Bootstrapped and tested on x86_64-unknown-linux-gnu.  I'm
double-checking because of the very high number of FAILs currently.

Richard.

	PR tree-optimization/112281
	* tree-loop-distribution.cc (pg_add_dependence_edges):
	Preserve stmt order when the innermost loop has exact
	overlap.

	* gcc.dg/torture/pr112281.c: New testcase.