c++: decltype of (non-captured variable) [PR83167]
Bootstrapped and regtested on x86_64-pc-linux-gnu, does this look OK for
trunk?

-- >8 --

For decltype((x)) within a lambda where x is not captured, we dubiously
require that the lambda has a capture default, unlike for decltype(x).
This patch fixes this inconsistency; I couldn't find a justification for
it in the standard.

	PR c++/83167

gcc/cp/ChangeLog:

	* semantics.cc (finish_decltype_type): If capture_decltype
	returns NULL_TREE, fall back to the ordinary code path.
	(capture_decltype): Return NULL_TREE if the lambda has no
	capture-default.

gcc/testsuite/ChangeLog:

	* g++.dg/cpp0x/lambda/lambda-decltype4.C: New test.